fixed it to send probabilities that are scaled correctly to the python file, now the finale of finding out what's wrong with this model

+++ EspIdfCode/main/PrintFunctions.h
@@ -12,6 +12,6 @@ bool ReadForReadiness();
 
 void CustomWrite(size_t number);
 
-void CustomWrite(int8_t number);
+void CustomWrite(float number);
 
 void CustomWrite(uint8_t* buf);
